requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
face-recognition>=1.3.0
pytesseract>=0.3.8
google-api-python-client>=2.0.0
twilio>=7.0.0
python-dotenv>=0.19.0
numpy>=1.19.0
Pillow>=8.0.0
gTTS>=2.2.0
playsound>=1.3.0
